\name{confReg}
\Rdversion{1.0}
\alias{confReg}
\alias{confReg-package}
\docType{package}

\title{Estimating Confidences of Individual Regression Predictions}
\description{
  This package contains various functions to estimate the confidences of individual predictions made with regression prediction. Therefore, regression algorithms and a simple feature selection routine is implemented.
 }


\details{
  \tabular{ll}{
    Package: \tab confReg\cr
    Type: \tab Package\cr
    Version: \tab 0.1.1\cr
    Date: \tab 2011-05-20\cr
    License: \tab GPL (>=2.0)\cr
    LazyLoad: \tab yes\cr
    Depends: \tab methods, class, e1071\cr
  }

  This package contains various functions to estimate the confidences of individual predictions
  made with regression. See also \emph{Briesemeister et al., No longer confidential: Estimating Confidences of Individual Regression Predictions, to be submitted}.
}


\references{
Briesemeister et al., No longer confidential: Estimating Confidences of Individual Regression Predictions, to be submitted
}
\author{
Sebastian Briesemeister
}


\keyword{package}
\keyword{documentation}